test_util: move csr test utilities into dedicated library
